refactor: types.tsを9つのドメイン別ファイルに分割 (Phase 1)
- domain/types/ 配下に world, player, enemy, gimmicks, items, stage,
  game-state, feedback, audio の9ファイルに型定義を分割
- domain/types/index.ts で barrel export を提供
- 既存 types.ts を barrel re-export に変更し後方互換性を維持
- 22項目の分割テスト (domain-types.test.ts) を追加
- tasks.md の Phase 1 チェックリストを完了に更新
