+++ src/features/ipne/domain/types/index.ts
@@ -1,0 +1,14 @@
+/**
+ * domain/types barrel export
+ * 全ドメイン型定義の再エクスポート
+ */
+
+export * from './world';
+export * from './player';
+export * from './enemy';
+export * from './gimmicks';
+export * from './items';
+export * from './stage';
+export * from './game-state';
+export * from './feedback';
+export * from './audio';
